feat(scrapper): add predefined resources
This commit introduces "predefined" submodule
which contains resource that are not easily
parsable from web, because their main
value is around the "planets" property
which usually is not represented structurally
on web, but instead in form of text.

To avoid using LLMs for this project,
those 22 (16 natural and 6 atmospheric) resources
were defined in the "predefined" module and
are used in fetchArticle function whenever
the url matches any of the predefined resources.

This commit also strips previously introduced
support for atmospheric resources in parsers.