feat(scrapper): introduce url composer and fetch deep functions
This commit introduces removes explicit references
to "fandom.com" and replaces them with urlComposer.
The use of fandom.com provider for fetching
is discouraged because of TOS through code comments
and will be mentioned in README later.

This commit also introduces fetchArticleDeep
function which recursively calls fetchArticle
function for each ingradient in root article
recipe and returns a tree, which shows
recipe for every single ingradient deeply.