feat(scrapper, types): introduce "cache" submodule
This commit introduces "cache" submodule in the
scrapper workspace package which defines
logic for caching, storing and retrieving articles
from cache. For this the fetchArticle
was updated to accept "strategy" option: "cache"
or "remote" (same as "no-cache").

Adding this submodule allows to improve fetching speeds
and reduce amount of calls to remote providers.

This commit also adds searchCachedArticles
function, which allows to search through
articles by their name or key.