feat(scrapper): implement super deep fetching
This commit extends fetchArticleDeep
function to output FULL tree of all ingradients. The part
in question is the "craftedAt" property. The function was
modded to include recipes for "craftedAt" machines as well.

This made the entire graph in fully expanded view for RTG
recipe to exceed any viewport limitations. The amount
of nodes for the RTG recipe was 152.