fix(web): disallow toggling child trees
This commit prevents the user from toggling
child article graph nodes and establishes that once
the article has been expanded (hidden=false), it cannot
be hidden again.

The main mindset of the graph is to show as many nodes as possible.
All the nodes are hidden by default to prevent resource exhaustion.