fix(scrapper): support uri encoded keys
This commit fixes the issue when fetchArticle
and cacheArticles didn't properly
handle encoded uri keys and
instead saved them as raw in the _cache.

This created inconsistency and potential
bugs when trying to fetch some of the items.

This commit fixes the bug by ensuring
that axios (in fetchHTML) performs the necessary
uri encoding when calling remote resource. And
getKeyFromURL did the necessary decoding
on the input when needed.

TDLR: getKeyFromURL now supports urls with
uri encoded and decoded article keys.