fix(scrapper): display correct ingredient amounts
This commit fixes the issue when sometimes ingredient
amounts would not be correct. The issue had to do
with missing support cases in parseNumber utility,
which didn't strip letters (with values like "x2" and "ABC4").

The utilized function - parseInt(#, 10) - does strip
some letters, but not all of them and order of them matters.
Therefore payload like "x2" would get parsed as NaN
and default to amount 1.

The bug was fixed by applying remove letters operation
to the "cut" stage in parseNumber utility.