feat(scrapper): optimize fetchArticleDeep function
This commit strips the fetchArticleDeep and defines
fetchArticleWithRefs as an alternative. The purpose
of those functions is still the same, but they
use different approaches when it comes to constructing
payload for a graph.

While fetchArticleDeep (:=> ArticleNode) produced
a Tree Node, where parent Node deeply contained information
about every single, fetchArticleWithRefs (:=> ArticleWithRefs)
produces a struct with article and "_referencesMap", where
article is the top level article, and "_referencesMap"
is a deeply resolved lookup table.

The main reason for this change is terrible
memory consumption of fetchArticleDeep method
and ArticleNode data structure.
It would duplicate the same articles if they
are referenced by different nested parent articles.
This was okay for cases when structure is constructed
and consumed in the same process. But once this
structure started getting utilized in HTTP communication,
it was very obvious how unperformant it is.

ArticleWithRefs instead stores all references
that the root article may need when resolving
deeply, removing duplication and providing
a lookup table ("_referencesMap") with O(1)
lookup time complexity.

This optimization brought down API packet
size for deep resolution of "RTG" article
from 10.5KB to 2.5KB (x4 improvement).