feat(scrapper): add utils for fetching article parents
This commit introduces the following functions:
- get-parents-for-cached-article
- fetch-parents-for-article
- fetch-parents-for-article-key
- fetch-parents-with-refs-for-article
- fetch-parents-with-refs-for-article-key

All of those utilities are centered around
fetching article parents and constructing
ArticleWithRefs struct on demand for traversing
through tree with bottom-to-top approach.

This is useful, when we're trying to visualize
dependency between child and its parents, while
child being the root node - this refers to "show
for articles that use this ingredient" feature.