feat(web): implement usages mode
This commit implements a new mode for
viewing articles. Previously only mode "recipe"
was availablle. With this commit new mode "usages"
can be used to utilize the "bottom-to-top" approach.
This allows to see which articles are utilizing the current
article (as opposed to how in "recipe" the user can see
which articles the current article utilizes - "top-to-bottom").