feat(web): add support for history
This commit implements support for articles history.
Whenever user opens an article, this will be reflected
under the search bar. This allows to quickly navigate
to pages you've been to before.

This solution doesn't include communication
persistent storage solution (like localStorage).
Which means that after refreshing the page
the history will be reset.