feat(web): allow to mark ingredients as done
This commit implements logic for marking article nodes
as done using a button on the bottom right side of the
node. Upon pressing on the button, the node will turn
green, indicating that is has been already collected/crafted.

This is a neat feature to simplify the process of keeping track of
what resources (ingredients) have to be collected for crafting.