fix(web): replace no usages badge with counter
This commit modifies the behaviour of the "show usages"
button when in "recipe" mode. Previously, if the user
would click on "expand parents" ("show usages"), an API call
would be performed to verify if there are any parents. If there
are none, then the button would be replaced with a badge saying
"No Usages". This was inconsistent with UI practices for this project.

Therefore the logic has been modified to keep the "show usages"
button, but show "0" in the counter badge and make the button
disabled if the backend responded with 0 parents.