feat(web): make articles history persistent
This commit implements persistent articles history.
Now, whenever the user refreshes or closes the application,
the state of the previously oppened articles will be saved
in localStorage. Upon the first render of the application,
the local storage will be prompted to return saved articles
history state, it will be assigned to articles history store
context value.