feat(web): replace history with pins
This commit replaces historical articles placed
below the search bar with pinned articles. The history
store has not been removed, and the content of it
will later be displayed in a sidebar, openable using
control buttons. Pins feature allows users to pin
one or multiple articles and have a quick access to them
at any point of time.